perf(web): defer personal grant display with scope controls
