+++ apps/web/src/components/personal-resource-attachment-control.tsx
@@ -1,3 +1,5 @@
+import { useEffect, useState } from "react";
+import { ongoingPersonalResourceNames } from "@/lib/personal-resource-ongoing-access";
 import { PersonalResourceScopeChoice } from "./personal-resource-scope-choice";
 import { RefreshCwIcon } from "lucide-react";
 
@@ -23,6 +25,7 @@ export function PersonalResourceAttachmentControl(props: {
     controller.notice !== null ||
     controller.error !== null ||
     controller.truncated;
+  const ongoingNames = useOngoingNames(controller);
   if (!controller.eligible || !hasVisibleStatus) {
     return null;
   }
@@ -33,10 +36,10 @@ export function PersonalResourceAttachmentControl(props: {
       className={cn("min-w-0 space-y-2", props.compact ? "mt-2" : "mt-4")}
       aria-busy={controller.loading || controller.refreshing}
     >
-      {controller.ongoingResourceNames.length > 0 ? (
+      {ongoingNames.length > 0 ? (
         <p className="text-xs text-fg-muted">
-          Existing ongoing authorization: {controller.ongoingResourceNames.join(", ")}. This is
-          separate from your next-message choice.
+          Existing ongoing authorization: {ongoingNames.join(", ")}. This is separate from your
+          next-message choice.
         </p>
       ) : null}
       {showScopeChoice ? (
@@ -81,4 +84,57 @@ export function PersonalResourceAttachmentControl(props: {
       ) : null}
     </div>
   );
+}
+
+function useOngoingNames(controller: PersonalResourceAttachmentController): string[] {
+  const [checkedAt, setCheckedAt] = useState(Date.now);
+  const authorities = controller.catalog
+    ? [
+        ...controller.catalog.variableSetAuthorities,
+        ...controller.catalog.rigAuthorities,
+        ...controller.catalog.connectedMachineAuthorities,
+      ]
+    : [];
+  const nextExpiry = Math.min(
+    ...authorities.flatMap((authority) =>
+      authority.grants.flatMap((grant) =>
+        grant.status === "active" &&
+        grant.expiresAt !== null &&
+        Date.parse(grant.expiresAt) > checkedAt
+          ? [Date.parse(grant.expiresAt)]
+          : [],
+      ),
+    ),
+  );
+  useEffect(() => {
+    if (!Number.isFinite(nextExpiry)) return;
+    const timer = setTimeout(
+      () => setCheckedAt(Date.now()),
+      Math.min(2_147_483_647, Math.max(1, nextExpiry - Date.now() + 1)),
+    );
+    return () => clearTimeout(timer);
+  }, [nextExpiry, checkedAt]);
+  if (!controller.ongoingScope) return [];
+  return ongoingPersonalResourceNames({
+    ...controller.ongoingScope,
+    authorities,
+    resources: [
+      ...controller.selected.variableSets.map((resource) => ({
+        kind: "variable_set" as const,
+        id: resource.id,
+        name: resource.name,
+      })),
+      ...controller.selected.rigs.map((resource) => ({
+        kind: "rig" as const,
+        id: resource.id,
+        name: resource.name,
+      })),
+      ...controller.selected.connectedMachines.map((resource) => ({
+        kind: "connected_machine" as const,
+        id: resource.enrollmentId,
+        name: resource.name,
+      })),
+    ],
+    now: Math.max(checkedAt, Date.now()),
+  });
 }
